refactor(Seating) clean up js modules
