+++ src/Tickets/Seating/app/service/iframe.js
@@ -1,5 +1,5 @@
 import { notifyUserOfError } from '@tec/tickets/seating/service/notices';
-import { establishReadiness } from './service-api';
+import { establishReadiness } from './api';
 
 const { _x } = wp.i18n;
 
@@ -68,7 +68,7 @@ window.tec = window.tec || {};
 window.tec.tickets.seating = window.tec.tickets.seating || {};
 window.tec.tickets.seating.service = window.tec.tickets.seating.service || {};
 window.tec.tickets.seating.service.iframe = {
-	...(window.tec.tickets.seating.serviceiframe || {}),
+	...(window.tec.tickets.seating.service.iframe || {}),
 	getIframeElement,
 	initServiceIframe,
 };
