+++ src/Tickets/Seating/app/service/index.js
@@ -1,4 +1,4 @@
 import './errors.js';
 import './notices.js';
-import './service-api';
+import './api';
 import './iframe.js';
